% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_kiwi_harvest.R
\name{max_kiwi_harvest}
\alias{max_kiwi_harvest}
\title{max kiwi harvest}
\usage{
max_kiwi_harvest(K, r)
}
\arguments{
\item{K}{farm capacity}

\item{r}{intrinsic growth rate}
}
\description{
max kiwi harvest
}
