% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{snacks calculator}
\usage{
snacks(grapes, apples, burritos)
}
\arguments{
\item{grapes}{mass of grapes in grams}

\item{apples}{number of apples in a bucket}

\item{burritos}{per pound}
}
\description{
snacks calculator
}
\examples{
snacks(grapes = 2.5, apples = 6, burritos = 1)
}
